refactor: extract browser options (#2654)
